Fix : devide home, chat area, connect websocket
